Attempt refactor to allow for `uv tool install`
