--- requirements.txt
@@ -1,8 +1,0 @@
-# DW3 Survey Logger - runtime dependencies
-# Stdlib modules (tkinter, sqlite3, etc.) are intentionally not listed.
-
-pynput>=1.7.6
-
-# Windows-only (used for optional window focus helpers)
-pywin32>=306; platform_system=="Windows"
-openpyxl>=3.1.2
