+++ src/dw3_survey_logger/journal_monitor.py
@@ -28,7 +28,7 @@ from collections import deque
 if TYPE_CHECKING:
     from journal_state_manager import JournalStateManager
 
-from observer_models import generate_event_id
+from .observer_models import generate_event_id
 
 import logging
 logger = logging.getLogger("dw3.journal_monitor")
