+++ src/dw3_survey_logger/observer_overlay.py
@@ -31,7 +31,7 @@ from pathlib import Path
 from typing import Optional, Dict, Any, Callable, Tuple
 from dataclasses import dataclass
 
-from observer_models import (
+from .observer_models import (
     ObserverNote,
     SliceStatus,
     SamplingMethod,
@@ -40,7 +40,7 @@ from observer_models import (
     create_observation_from_context,
     SURVEY_AXIS_INDEX,
 )
-from journal_state_manager import CurrentContext
+from .journal_state_manager import CurrentContext
 
 import logging
 logger = logging.getLogger("dw3.observer_overlay")
